Add working popup-window for timetableclick
