+++ src/dev/components/user/CourseInfo.jsx
@@ -1,10 +1,30 @@
 import React from 'react'
 import { bindActionCreators } from 'redux'
 import { connect } from 'react-redux'
+import axios from 'axios'
 
 import * as actionCreators from '../../actions/courses.js'
 
 class CourseInfo extends React.Component {
+
+  makeReservation() {
+    var JOOGAURL = typeof(JOOGASERVER) === "undefined" ? 'http://localhost:3000/reserveSlot' : JOOGASERVER+'/reserveSlot'
+    console.log("JOOGASERVER: ", JOOGASERVER);
+    console.log("JOOGAURL: ", JOOGAURL);
+    var that = this;
+    console.log("CLICK", this.props);
+    axios.post(
+      JOOGAURL, {
+        user: this.props.currentUser.key,
+        slot: this.props.item
+      })
+      .then( response => {
+        console.log(response);
+      })
+      .catch( error => {
+        console.log(error);
+      });
+  }
 
   exitContainer() {
     this.props.actions.removeCourseInfo()
@@ -15,17 +35,19 @@ class CourseInfo extends React.Component {
     if(this.props.courseInfo) {
       return (
         <div className="course-info-container">
-          <button className="exit-btn" onClick={this.exitContainer.bind(this)}>x</button>
-          <h3>{this.props.courseInfo.courseType.name}</h3>
-          <p>{this.props.courseInfo.courseType.desc}</p>
-          <hr/>
-          <p>Klo {this.props.courseInfo.start} - {this.props.courseInfo.end}</p>
-          <p>Sijainti {this.props.courseInfo.place.name}, {this.props.courseInfo.place.address}</p>
-          <hr/>
-          <p>Joogaopettaja {this.props.courseInfo.instructor.name}</p>
-          <hr/>
-          <p>Ilmoittautuneita {this.props.courseInfo.users.length}/{this.props.courseInfo.maxCapacity}</p>
-          <button className="btn-small">Ilmoittaudu</button>
+          <div className="course-info">
+            <button className="exit-btn" onClick={this.exitContainer.bind(this)}>x</button>
+            <h3>{this.props.courseInfo.courseType.name}</h3>
+            <p>{this.props.courseInfo.courseType.desc}</p>
+            <hr/>
+            <p>Klo {this.props.courseInfo.start} - {this.props.courseInfo.end}</p>
+            <p>Sijainti {this.props.courseInfo.place.name}, {this.props.courseInfo.place.address}</p>
+            <hr/>
+            <p>Joogaopettaja {this.props.courseInfo.instructor.name}</p>
+            <hr/>
+            <p>Ilmoittautuneita {this.props.courseInfo.users.length}/{this.props.courseInfo.maxCapacity}</p>
+            <button className="btn-small" onClick={this.makeReservation.bind(this)} >Ilmoittaudu</button>
+          </div>
         </div>
       )
     } else {
@@ -37,7 +59,7 @@ class CourseInfo extends React.Component {
 }
 
 function mapStateToProps(state) {
-  return { courseInfo: state.courseInfo }
+  return { courseInfo: state.courseInfo, currentUser: state.currentUser }
 }
 
 function mapDispatchToProps(dispatch) {
