+++ src/styles/app.scss
@@ -243,3 +243,38 @@ form {
     color: red;
   }
 }
+
+.course-info-container {
+  position: fixed;
+  z-index: 1;
+  left: 0;
+  top: 0;
+  width: 100%;
+  height: 100%;
+  overflow: auto;
+  background-color: rgba(0,0,0,0.4);
+  .course-info {
+    background-color: #fefefe;
+    margin: 15% auto;
+    padding: 20px;
+    border: 1px solid #888;
+    width: 80%;
+    border-radius: 3px;
+    box-shadow: 0 3px 10px #444;
+    .exit-btn {
+      color: #ff2233;
+      float: right;
+      font-size: 28px;
+      font-weight: bold;
+      padding: 0;
+      margin: 0;
+      line-height: 0;
+      border: none;
+      background-color: #fefefe;
+      border-radius: 3px;
+      &:hover {
+        color: #444;
+      }
+    }
+  }
+}
